FIXED: Use 64-bit millisecond timer
Now uses SDL_GetTick() for portability and consistently type as `int64_t`.
This avoids 49 days overflow on Windows.

+++ src/h/graphics.h
@@ -843,7 +843,7 @@ NewClass(eventobj)
   Int		x;			/* X coordinate relative to window */
   Int		y;			/* Y coordinate relative to window */
   Point		position;		/* Computed Position */
-  unsigned long time;			/* Time of event in milliseconds */
+  uintptr_t	time;			/* Time of event in milliseconds */
 End;
